requirements.txt
fastapi
uvicorn
streamlit
requests
python-multipart
google-cloud-texttospeech
python-pptx
